chore: Breaks the dataViewCommon folder into TableCollection and Pagination (#17132)
