+++ superset-frontend/src/components/Pagination/Wrapper.test.tsx
@@ -19,7 +19,7 @@
 
 import React from 'react';
 import { render, screen } from 'spec/helpers/testing-library';
-import Pagination from '.';
+import Wrapper from './Wrapper';
 
 jest.mock('./Next', () => ({
   Next: () => <div data-test="next" />,
@@ -36,34 +36,34 @@ jest.mock('./Ellipsis', () => ({
 
 test('Pagination rendering correctly', () => {
   render(
-    <Pagination>
+    <Wrapper>
       <li data-test="test" />
-    </Pagination>,
+    </Wrapper>,
   );
   expect(screen.getByRole('navigation')).toBeInTheDocument();
   expect(screen.getByTestId('test')).toBeInTheDocument();
 });
 
 test('Next attribute', () => {
-  render(<Pagination.Next onClick={jest.fn()} />);
+  render(<Wrapper.Next onClick={jest.fn()} />);
   expect(screen.getByTestId('next')).toBeInTheDocument();
 });
 
 test('Prev attribute', () => {
-  render(<Pagination.Next onClick={jest.fn()} />);
+  render(<Wrapper.Next onClick={jest.fn()} />);
   expect(screen.getByTestId('next')).toBeInTheDocument();
 });
 
 test('Item attribute', () => {
   render(
-    <Pagination.Item onClick={jest.fn()}>
+    <Wrapper.Item onClick={jest.fn()}>
       <></>
-    </Pagination.Item>,
+    </Wrapper.Item>,
   );
   expect(screen.getByTestId('item')).toBeInTheDocument();
 });
 
 test('Ellipsis attribute', () => {
-  render(<Pagination.Ellipsis onClick={jest.fn()} />);
+  render(<Wrapper.Ellipsis onClick={jest.fn()} />);
   expect(screen.getByTestId('ellipsis')).toBeInTheDocument();
 });
